added themed icon font color and size

+++ src/system/theme_manager/theme_manager.h
@@ -19,53 +19,53 @@ extern struct theme_data_t {
     // idle screen status area
     lv_color_t idle_screen_status_area_font_color;
     lv_color_t idle_screen_status_area_font_outline;
-    short int idle_screen_status_area_font_size;
+    uint8_t idle_screen_status_area_font_size;
     
     // idle screen softkey area
     lv_color_t idle_screen_softkey_area_font_color;
     lv_color_t idle_screen_softkey_area_font_outline;
-    short int idle_screen_softkey_area_font_size;
+    uint8_t idle_screen_softkey_area_font_size;
     lv_color_t idle_screen_softkey_area_bg_color;
     char idle_screen_softkey_area_bg_source[128];
     
     // idle screen wallpaper
     char idle_screen_wallpaper_source[128];
     lv_color_t idle_screen_wallpaper_color;
-    short int idle_screen_wallpaper_transparency;  // 0-255
+    uint8_t idle_screen_wallpaper_transparency;  // 0-255
 
     // menu app status area
     lv_color_t menu_app_status_area_font_color;
     lv_color_t menu_app_status_area_font_outline;
-    short int menu_app_status_area_font_size;
+    uint8_t menu_app_status_area_font_size;
     
     // menu app softkey area
     lv_color_t menu_app_softkey_area_font_color;
     lv_color_t menu_app_softkey_area_font_outline;
-    short int menu_app_softkey_area_font_size;
+    uint8_t menu_app_softkey_area_font_size;
     lv_color_t menu_app_softkey_area_bg_color;
     char menu_app_softkey_area_bg_source[128];
     
     // menu app wallpaper
     char menu_app_wallpaper_source[128];
     lv_color_t menu_app_wallpaper_color;
-    short int menu_app_wallpaper_transparency;  // 0-255
+    uint8_t menu_app_wallpaper_transparency;  // 0-255
 
     // menu app grid menu unselected
     lv_color_t menu_app_grid_menu_unselected_font_color;
     lv_color_t menu_app_grid_menu_unselected_font_outline;
-    short int menu_app_grid_menu_unselected_font_size;
+    uint8_t menu_app_grid_menu_unselected_font_size;
 
     // menu app grid menu selected
     lv_color_t menu_app_grid_menu_selected_font_color;
     lv_color_t menu_app_grid_menu_selected_font_outline;
-    short int menu_app_grid_menu_selected_font_size;
+    uint8_t menu_app_grid_menu_selected_font_size;
     lv_color_t menu_app_grid_menu_selected_bg_color;
     char menu_app_grid_menu_selected_bg_source[128];
-    short int menu_app_grid_menu_selected_transparency;  // 0-255
+    uint8_t menu_app_grid_menu_selected_transparency;  // 0-255
     
     // menu app grid menu icon 
-    short int menu_app_grid_menu_icon_width;
-    short int menu_app_grid_menu_icon_height;
+    uint8_t menu_app_grid_menu_icon_width;
+    uint8_t menu_app_grid_menu_icon_height;
     
     // Future fields (placeholders)
     // int grid_rows;
